Added sampling of the Loss Derivative function
GD misbehaves around points where the Loss function is not smooth.
A workaround it to make several samplings of the Loss Derivative function
with varying delta steps and average them. This essentially approximates
a smooth function around that point if the function value change is not huge.

The downside of this approach is that it slows GD a lot an requires
setting fewer GD iterations per frame.

Anyway, the workaround noticeably reduces jittering of the effectors.